Add embedded association conflict regressions
